backend/pyproject.toml
[tool.poetry]
name = "dans-utils"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.22"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.0"
pandas = "^2.1.2"
tabulate = "^0.9.0"
requests = "^2.31.0"
celery = "^5.3.4"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
beautifulsoup4 = "^4.12.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pytest = "^7.4.3"
httpx = "^0.25.1"
pydantic = {extras = ["email"], version = "^2.4.2"}
watchdog = "^3.0.0"
aiohttp = "^3.8.6"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
datetime = "^5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
